Refactor editor and agent structure
